Move colors stuff in jupytergis-core
